Swapped from using effect_raw to effect from the GA API
This was annoying...but apparently for some cards the API doesn't send the raw effect text.

In order to display effect text in app I used the version that HTML span tags.

Using a recursive regex I remove all occurrences of <span> turning the string into a "raw" version and then displaying that.