More Player model integration.
Removed increment and decrement tap history properties from PlayerCounterView because they're useless.

Updated the CounterButtonView to pass along value changes to the PlayerCounterView so that they can be tracked and saved in the player model.

App now saves game state when backgrounding or quitting.

Updated the Player model to conform to Codable.